requirements.txt
langchain>=0.1.0
faiss-cpu>=1.7.4
pypdf2>=3.0.0
streamlit>=1.25.0
python-dotenv>=1.0.0
langchain_groq>=0.1.0 
langchain-community>=0.0.21
langchain-huggingface
sentence-transformers